wrote Tests for the lean subsystem and added a package for the lean test in the test package
