--- Competition2018/src/test/java/competition/subsystems/lean/LeanLeftCommandTest.java
@@ -1,36 +1,0 @@
-package competition.subsystems.lean;
-
-import static org.junit.Assert.assertEquals;
-import org.junit.Test;
-
-import competition.BaseCompetitionTest;
-import competition.subsystems.lean.LeanSubsystem;
-import competition.subsystems.lean.commands.LeanLeftCommand;
-
-public class LeanLeftCommandTest extends BaseCompetitionTest {
-
-    LeanSubsystem lean;
-    LeanLeftCommand command;
-
-    @Override
-    public void setUp() {
-        super.setUp();
-
-        lean = injector.getInstance(LeanSubsystem.class);
-        command = injector.getInstance(LeanLeftCommand.class);
-    }
-
-    @Test
-    public void simpleTest() {
-        command.initialize();
-        command.execute();
-    }
-
-    @Test
-    public void checkLeanLeft() {
-        assertEquals(0.0, lean.motor.getMotorOutputPercent(), 0.001);
-        command.initialize();
-        command.execute();
-        assertEquals(0.4, lean.motor.getMotorOutputPercent(), 0.001);
-    }
-}
