+++ Competition2018/src/test/java/competition/subsystems/lean/commands/DecreaseSpeedCommandTest.java
@@ -1,23 +1,23 @@
-package competition.subsystems.lean;
+package competition.subsystems.lean.commands;
 
 import static org.junit.Assert.assertEquals;
 import org.junit.Test;
 
 import competition.BaseCompetitionTest;
 import competition.subsystems.lean.LeanSubsystem;
-import competition.subsystems.lean.commands.StopLeaningCommand;
+import competition.subsystems.lean.commands.DecreaseSpeedCommand;
 
-public class StopLeaningCommandTest extends BaseCompetitionTest {
+public class DecreaseSpeedCommandTest extends BaseCompetitionTest {
 
     LeanSubsystem lean;
-    StopLeaningCommand command;
+    DecreaseSpeedCommand command;
 
     @Override
     public void setUp() {
         super.setUp();
 
         lean = injector.getInstance(LeanSubsystem.class);
-        command = injector.getInstance(StopLeaningCommand.class);
+        command = injector.getInstance(DecreaseSpeedCommand.class);
     }
 
     @Test
@@ -27,11 +27,12 @@ public class StopLeaningCommandTest extends BaseCompetitionTest {
     }
 
     @Test
-    public void checkStopLeaning() {
+    public void checkDecreaseSpeed() {
         lean.leanRight();
         assertEquals(-0.4, lean.motor.getMotorOutputPercent(), 0.001);
         command.initialize();
         command.execute();
-        assertEquals(0.0, lean.motor.getMotorOutputPercent(), 0.001);
+        lean.leanRight();
+        assertEquals(-0.1, lean.motor.getMotorOutputPercent(), 0.001);
     }
 }
