+++ Competition2018/src/test/java/competition/subsystems/lean/commands/LeanRightCommandTest.java
@@ -1,23 +1,23 @@
-package competition.subsystems.lean;
+package competition.subsystems.lean.commands;
 
 import static org.junit.Assert.assertEquals;
 import org.junit.Test;
 
 import competition.BaseCompetitionTest;
 import competition.subsystems.lean.LeanSubsystem;
-import competition.subsystems.lean.commands.IncreaseSpeedCommand;
+import competition.subsystems.lean.commands.LeanRightCommand;
 
-public class IncreaseSpeedCommandTest extends BaseCompetitionTest {
+public class LeanRightCommandTest extends BaseCompetitionTest {
 
     LeanSubsystem lean;
-    IncreaseSpeedCommand command;
+    LeanRightCommand command;
 
     @Override
     public void setUp() {
         super.setUp();
 
         lean = injector.getInstance(LeanSubsystem.class);
-        command = injector.getInstance(IncreaseSpeedCommand.class);
+        command = injector.getInstance(LeanRightCommand.class);
     }
 
     @Test
@@ -27,13 +27,10 @@ public class IncreaseSpeedCommandTest extends BaseCompetitionTest {
     }
 
     @Test
-    public void checkIncreaseSpeed() {
-        lean.decreaseSpeed();
-        lean.leanRight();
-        assertEquals(-0.1, lean.motor.getMotorOutputPercent(), 0.001);
+    public void checkLeanRight() {
+        assertEquals(0.0, lean.motor.getMotorOutputPercent(), 0.001);
         command.initialize();
         command.execute();
-        lean.leanRight();
         assertEquals(-0.4, lean.motor.getMotorOutputPercent(), 0.001);
     }
 }
